* Fixing `parse_config` for gphoto2 * Added `is_sleeping` for pocs. * Dockerfile install local via pip.
