+++ scripts/testing/test-software.sh
@@ -1,12 +1,12 @@
-#!/bin/bash -e
+#!/usr/bin/env bash
 
 clear;
 
-cat << EOF
+cat <<EOF
 Beginning test of pocs software. This software is run inside a virtualized docker
 container that has all of the required dependencies installed.
 
-This will start a single docker container, mapping the host $PANDIR into the running docker
+This will start a single docker container, mapping the host ${PANDIR} into the running docker
 container, which allows for testing of any local changes.
 
 You can view the output for the tests in a separate terminal:
@@ -22,8 +22,8 @@ sleep 5;
 
 docker run --rm -it \
     -e LOCAL_USER_ID=$(id -u) \
-    -v /var/panoptes/POCS:/var/panoptes/POCS \
-    -v /var/panoptes/logs:/var/panoptes/logs \
+    -v "${POCS}":/var/panoptes/POCS \
+    -v "${PANDIR}"/logs:/var/panoptes/logs \
     pocs:testing \
     "/var/panoptes/POCS/scripts/testing/run-tests.sh"
