+++ setup.cfg
@@ -42,11 +42,8 @@ install_requires =
     astroplan
     astropy
     Flask
-    gcloud
-    google-cloud-storage
     matplotlib
     numpy
-    pandas
     panoptes-utils>=0.2.17
     pyserial>=3.1.1
     pendulum
@@ -78,8 +75,9 @@ testing =
     responses
     coverage
     pytest-remotedata>=0.3.1'
-social =
-    tweepy
+google =
+    gcloud
+    google-cloud-storage
 
 [options.entry_points]
 # Add here console scripts like:
